Server side tests -- exclude any that need to use a db connection
