+++ .github/workflows/main.yml
@@ -56,5 +56,5 @@ jobs:
         working-directory: server
       - run: npm run lint -- --no-fix
         working-directory: server
-      - run: npm run test
+      - run: npm run test:server
         working-directory: server
