+++ server/package.json
@@ -36,7 +36,7 @@
     "test:one": "NODE_ENV=test DEBUG='testing, server, error' node_modules/.bin/mocha --require @babel/register --require @babel/runtime/regenerator --require @babel/polyfill src/mcr/activity/activity.server.spec.js",
     "test:one:demo": "NODE_ENV=test DEBUG='server, error' node_modules/.bin/mocha --require @babel/register --require @babel/runtime/regenerator --require @babel/polyfill src/mcr/demo/demo.server.spec.js",
     "test:server": "npm run build && node_modules/.bin/mocha --recursive --sort --bail  'build/**/*.server.spec.js'",
-    "test:system": "npm run build && node_modules/.bin/mocha --recursive --sort --bail 'build/**/*.system.spec.js'",
+    "test:system": "npm run build && NODE_ENV=test node_modules/.bin/mocha  dotenv/config --recursive --sort --bail 'build/**/*.system.spec.js'",
     "test:verbose": "npm run build && NODE_ENV=test DEBUG='server, error' node_modules/.bin/mocha dotenv/config --recursive --sort --bail 'build/**/*.spec.js'",
     "test:verbose.comment": "make test.versbose the same as test but add in DEBUG"
   },
