+++ server/src/mcr/assignment/assignment.system.spec.js
@@ -1,0 +1,9 @@
+var should = require('should')
+const mongoose = require('mongoose')
+import Helper from '../common/test-helper'
+const helper = new Helper()
+import Model from './assignment'
+
+const typeName = 'Assignment'
+
+/* global describe it */
